feat: outlet layout component for conditional header, footer
