+++ src/index.jsx
@@ -3,7 +3,6 @@ import ReactDOM from "react-dom/client";
 import App from "./pages/App";
 import "./styles/reset.css";
 import "antd/dist/reset.css";
-import Layout from "./components/Layout";
 import GlobalStyle from "./styles/globalStyle";
 import { ThemeProvider } from "styled-components";
 import theme from "./styles/theme";
@@ -14,9 +13,7 @@ root.render(
     <>
         <ThemeProvider theme={theme}>
             <GlobalStyle />
-            <Layout>
-                <App />
-            </Layout>
+            <App />
         </ThemeProvider>
     </>,
 );
